refactor(use-database): rename useDatabase hook to useProductsList
Move reducer to hooks/useProductsList/
Rename reducer file to reducer.js
Move action handlers to handlers.js

Experimenting with ducks inspired useReducer organization
