+++ src/hooks/useProductsList/utils/createRandomProducts.js
