+++ src/hooks/useProductsList/utils/useConfigAutosave.js
